don't add margins if computedWidth of slide is zero
If an img hasn't loaded yet, outerWidth() will be zero. The marginLeft in that case would be half the container width. This breaks layout pretty badly when the image loads after adjustSlide() has already run. Better to not apply margins in this case, and listen for img load in another way.