Use <button> for keyboard accessibility
I'm all for having gotten away from <a href="#"> for the pager markup.
That's an accessibility anti-pattern. Anchor hrefs should only point to
external content, not a dummy placeholder.

But elements that have a more-or-less fixed role in loading in-page
content should really just be <button>s. This allows them to default to
conventional CSS focus outlines, and to be sequenced normally in the
tabindex for keyboard-only, non-mouse users. Buttons have much better
out-of-the-box accessibility for elements without ARIA roles:

http://yahoodevelopers.tumblr.com/post/59489724815/easy-fixes-to-common-
accessibility-problems

Also on a humourous, ranty note:

http://www.karlgroves.com/2013/05/14/links-are-not-buttons-neither-are-d
ivs-and-spans/

P.S: I smiled to discover that Builtwith says that cycle (1 & 2
combined) is used on more than 3 million sites. More than Modernizr!
Wow, congrats!