Fixed fx for manual progressive slideshows
The Fx of my manual progressive slideshow wasn't working the way I expected. The fx started on the old slide (slide to left or fade-out) and then briefly the background was visible and then the fx started on the new slide. The slides weren't synchronized anymore. A small timeout fixed this issue.

You can see the bug in the demo of the progressive loading manual slide:
http://jquery.malsup.com/cycle2/demo/progressive.php#manual